+++ tests/funcflow/loggers/test_json_logger.py
@@ -434,7 +434,7 @@ class TestMisc(unittest.TestCase):
         logger = PassThroughJsonLogger('default')
         log = logger.debug(lambda *_: 'msg')
         with self.assertRaises(AttributeError):
-            _ = pickle.dumps(log)
+            _ = pickle.loads(pickle.dumps(log))
 
 
 if __name__ == '__main__':