+++ tests/funcflow/loggers/test_std_logger.py
@@ -414,7 +414,7 @@ class TestMisc(unittest.TestCase):
         logger = PassThroughStdLogger('default')
         log = logger.debug(lambda *_: 'msg')
         with self.assertRaises(AttributeError):
-            _ = pickle.dumps(log)
+            _ = pickle.loads(pickle.dumps(log))
 
 
 if __name__ == '__main__':