+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+- Added Fallback to funcflow
+- Homogenized funcflow behavior on functions returning one-tuples
+- Added missing `__hash__` magic methods
+
 ## [0.5.2] - 2025-06-12
 - Fixed JSON logging issue with JsonObject
 - Added or-operator ("|") merging to JsonObject